+++ packages/tideline/test/index.js
@@ -20,21 +20,16 @@ import moment from 'moment-timezone'
 
 /* js/data/ */
 import './format.test'
-import './datetime.test'
 import './constants.test'
-import './basalutil.test'
 import './categorize.test'
 
 /* js/plot/ */
-import './annotations.test'
 import './commonbolus.test'
-import './device.test'
 
 // DOM required
 // ====================================
 
 /* plugins/ */
-import './basics_classifiers.test'
 import './basics_datamunger.test'
 
 import './blip/components/logic/actions.test.js'
@@ -42,6 +37,5 @@ import './blip/components/sitechange/Selector.test.js'
 import './blip/components/BasicsUtils.test.js'
 
 window.d3 = require('d3')
-window.d3.chart = require('d3.chart')
 
 moment.locale('en')