+++ CHANGELOG.md
@@ -10,10 +10,11 @@ It is based on Tidepool Blip 1.27.
 - YLP-2476 Use tide-whisperer v2: route /all.
 - YLP-2696 Fix some Sonarcloud bugs and code smells
 - YLP-2697 App routing should reveal data dependencies
-- YLP-2697 App routing should reveal data dependencies
 
 ### Changed
 - YLP-2574 Create tool tips independently of the name of the columns
+- YLP-2704 Remove settings logic from viz
+- YLP-2726 Cleanup tideline/viz packages
 
 ## 3.0.5-beta - 2023-07-12
 ### Added